+++ indico/core/signals/event/designer.py
@@ -11,5 +11,7 @@ from blinker import Namespace
 _signals = Namespace()
 
 print_badge_template = _signals.signal('print-badge-template', """
-Called when printing a badge template. The registration form is passed in the `regform` kwarg.
+Called when printing a badge template.
+The registration form is passed in the `regform` kwarg, the list of registration
+objects are passed in the `registrations` kwarg.
 """)